Fix Q16.16 bugs and adapt test suite for fixed-point
Core fixes:
- Rewrite kulua_str2number to apply exponent before Q16.16 conversion,
  fixing precision loss in scientific notation (e.g. .2e2 == 20)
- Reject malformed exponents ('1e', '0x1p+') in str2number/strx2number
- Fix kulua_ceil for negative numbers (was floor, now truncates toward zero)
- Fix kulua_fmod sign handling (truncate toward zero, not floor)
- Fix tostringbuffFloat uint32 overflow for fractional parts > 0.65
- Fix math.atan default argument (1 vs KULUA_ONE)
- Fix math.log base division (use Q16.16 division, not plain C division)
- Fix unsigned arithmetic in luai_numadd/luai_numsub/luai_numunm
- Fix kulua_int2num saturation (check bounds before multiply)
- Fix division by zero in luai_numdiv (return max/min instead of trap)
- Fix left-shift UB in kulua_fixed.c (use multiply instead of shift)
- Fix lstrlib.c pack/unpack saturation for float/double conversions

Test suite adaptation:
- Add _fixedpoint flag (auto-detected via math.huge < 100000)
- Guard ~40 test sections across 12 files for Q16.16 limitations:
  overflow range, NaN/inf absence, minint+0.0 saturation, precision
- Relax eq() tolerance to 0.05 for Q16.16 (vs 1E-5 for 32-bit float)
- Create testes/fixed.lua with Q16.16-specific assertions

All 34 test files pass: zig build && cd testes && lua -W -e "_U=true" all.lua

+++ lobject.c
@@ -445,7 +445,7 @@ static int tostringbuffFloat (lua_Number n, char *buff) {
   len += snprintf(buff + len, (size_t)(LUA_N2SBUFFSZ - len), "%d", (int)ipart);
   if (fpart != 0) {
     /* Convert fractional part: fpart/65536 as up to 5 decimal digits */
-    uint32_t f = ((uint32_t)fpart * 100000u + 32768u) / 65536u;
+    uint32_t f = (uint32_t)(((uint64_t)fpart * 100000u + 32768u) / 65536u);
     /* Remove trailing zeros */
     int ndigits = 5;
     while (ndigits > 1 && f % 10 == 0) { f /= 10; ndigits--; }
